test(volume): share the square-footprint lattice between both volume suites
The oracle-consensus volume test copied the latticeCloud helper (plus A_HALF and
the Field type) verbatim from analyticVolumeOracle.test.ts, which the duplication
gate counted as duplicated new code. Extract them to
tests/support/volumeLatticeFixture.ts and import from both, so the lattice lives
in one place and neither suite restates it.
